Fix implicit type conversions

+++ src/fheroes2/world/world_loadmap.cpp
@@ -180,7 +180,7 @@ bool World::LoadMapMP2( const std::string & filename, const bool isOriginalMp2Fi
 
     const int32_t worldSize = width * height;
 
-    if ( totalFileSize < MP2::MP2_MAP_INFO_SIZE + worldSize * MP2::MP2_TILE_STRUCTURE_SIZE + MP2::MP2_ADDON_COUNT_SIZE ) {
+    if ( totalFileSize < static_cast<size_t>( MP2::MP2_MAP_INFO_SIZE + worldSize * MP2::MP2_TILE_STRUCTURE_SIZE + MP2::MP2_ADDON_COUNT_SIZE ) ) {
         DEBUG_LOG( DBG_GAME, DBG_WARN, "Map file " << filename.c_str() << " is corrupted" )
         return false;
     }
@@ -192,7 +192,8 @@ bool World::LoadMapMP2( const std::string & filename, const bool isOriginalMp2Fi
     const uint32_t addonCount = fs.getLE32();
     std::vector<MP2::mp2addon_t> vec_mp2addons( addonCount );
 
-    if ( totalFileSize < MP2::MP2_MAP_INFO_SIZE + worldSize * MP2::MP2_TILE_STRUCTURE_SIZE + addonCount * MP2::MP2_ADDON_STRUCTURE_SIZE + MP2::MP2_ADDON_COUNT_SIZE ) {
+    if ( totalFileSize < static_cast<size_t>( MP2::MP2_MAP_INFO_SIZE + worldSize * MP2::MP2_TILE_STRUCTURE_SIZE + addonCount * MP2::MP2_ADDON_STRUCTURE_SIZE
+                + MP2::MP2_ADDON_COUNT_SIZE ) ) {
         DEBUG_LOG( DBG_GAME, DBG_WARN, "Map file " << filename.c_str() << " is corrupted" )
         return false;
     }
@@ -202,7 +203,8 @@ bool World::LoadMapMP2( const std::string & filename, const bool isOriginalMp2Fi
     }
 
     // If this assertion blows up it means that we are not reading the data properly from the file.
-    assert( fs.tell() == MP2::MP2_MAP_INFO_SIZE + worldSize * MP2::MP2_TILE_STRUCTURE_SIZE + addonCount * MP2::MP2_ADDON_STRUCTURE_SIZE + MP2::MP2_ADDON_COUNT_SIZE );
+    assert( fs.tell() == static_cast<size_t>( MP2::MP2_MAP_INFO_SIZE + worldSize * MP2::MP2_TILE_STRUCTURE_SIZE + addonCount * MP2::MP2_ADDON_STRUCTURE_SIZE
+                    + MP2::MP2_ADDON_COUNT_SIZE ) );
     const size_t afterAddonInfoPos = fs.tell();
 
     // Come back to the end of map info section and read information about MP2 tiles.
@@ -282,12 +284,12 @@ bool World::LoadMapMP2( const std::string & filename, const bool isOriginalMp2Fi
     }
 
     // If this assertion blows up it means that we are not reading the data properly from the file.
-    assert( fs.tell() == MP2::MP2_MAP_INFO_SIZE + worldSize * MP2::MP2_TILE_STRUCTURE_SIZE );
+    assert( fs.tell() == static_cast<size_t>( MP2::MP2_MAP_INFO_SIZE + worldSize * MP2::MP2_TILE_STRUCTURE_SIZE ) );
 
     // Go back to the section after the add-on structure information and read the rest of data.
     fs.seek( afterAddonInfoPos );
 
-    if ( totalFileSize < afterAddonInfoPos + MP2::MP2_CASTLE_COUNT * MP2::MP2_CASTLE_POSITION_SIZE ) {
+    if ( totalFileSize < afterAddonInfoPos + static_cast<size_t>( MP2::MP2_CASTLE_COUNT * MP2::MP2_CASTLE_POSITION_SIZE ) ) {
         DEBUG_LOG( DBG_GAME, DBG_WARN, "Map file " << filename.c_str() << " is corrupted" )
         return false;
     }
@@ -345,10 +347,10 @@ bool World::LoadMapMP2( const std::string & filename, const bool isOriginalMp2Fi
     }
 
     // If this assertion blows up it means that we are not reading the data properly from the file.
-    assert( fs.tell() == afterAddonInfoPos + MP2::MP2_CASTLE_COUNT * MP2::MP2_CASTLE_POSITION_SIZE );
+    assert( fs.tell() == afterAddonInfoPos + static_cast<size_t>( MP2::MP2_CASTLE_COUNT * MP2::MP2_CASTLE_POSITION_SIZE ) );
 
     if ( totalFileSize
-         < afterAddonInfoPos + MP2::MP2_CASTLE_COUNT * MP2::MP2_CASTLE_POSITION_SIZE + MP2::MP2_CAPTURE_OBJECT_COUNT * MP2::MP2_CAPTURE_OBJECT_POSITION_SIZE ) {
+         < afterAddonInfoPos + static_cast<size_t>( MP2::MP2_CASTLE_COUNT * MP2::MP2_CASTLE_POSITION_SIZE + MP2::MP2_CAPTURE_OBJECT_COUNT * MP2::MP2_CAPTURE_OBJECT_POSITION_SIZE ) ) {
         DEBUG_LOG( DBG_GAME, DBG_WARN, "Map file " << filename.c_str() << " is corrupted" )
         return false;
     }
@@ -398,7 +400,7 @@ bool World::LoadMapMP2( const std::string & filename, const bool isOriginalMp2Fi
 
     // If this assertion blows up it means that we are not reading the data properly from the file.
     assert( fs.tell()
-            == afterAddonInfoPos + MP2::MP2_CASTLE_COUNT * MP2::MP2_CASTLE_POSITION_SIZE + MP2::MP2_CAPTURE_OBJECT_COUNT * MP2::MP2_CAPTURE_OBJECT_POSITION_SIZE );
+            == afterAddonInfoPos + static_cast<size_t>( MP2::MP2_CASTLE_COUNT * MP2::MP2_CASTLE_POSITION_SIZE + MP2::MP2_CAPTURE_OBJECT_COUNT * MP2::MP2_CAPTURE_OBJECT_POSITION_SIZE ) );
 
     // TODO: find a way to use this value properly.
     const uint8_t obeliskCount = fs.get();
